refactor(agent-playground): extract helpers + types to subfolder
